Fix: Unit tests- but with errors
